\name{Barlow.Proschan}
\alias{Barlow.Proschan}
\docType{data}
\title{
    Part data for Kettelle Spares Allocation Example
}
\description{
    Dataset of four parts with Part ID (Unit), demand rate (Lam), repair time (TAT), and Cost (C)
}
\usage{Barlow.Proschan}
\format{
    An unsorted dataframe with 4 rows, with columns \code{$Unit}
    (ID number of the part), \code{$Lam} (Demand rate for spare part), \code{$TAT} (Repair, or turnaround, time for part)
	and \code{C} (Purchase cost part)
}
\source{
    Richard E. Barlow and Frank Proschan, \emph{Statistical Theory of Reliability and Life Testing}, Probability Models
    (Barlow & Proschan, 1975)
}
%\references{
%Richard E. Barlow and Frank Proschan,  \emph{Statistical Theory of Reliability and Life Testing, Probability Models} (Barlow & Proschan, 1975)
%}
\keyword{datasets}