\name{de.Sousa.Borges.METRIC}
\alias{de.Sousa.Borges.METRIC}
\docType{data}
\title{
    Part data for METRIC, two eschelon, example with more than one part
}
\description{
    Dataset of two parts handled at two bases and one depot
}
\usage{Barlow.Proschan}
\format{
    An dataframe with 4 rows, used as unsorted with columns \code{$Base} \code{$LRU}
    (Line Replaceable Unit), \code{$bLam} (Demand rate for spare part at base), \code{$bfT} (Repair time at base),
	\code{$Pbr} (Probability of base repair), \code{$transp} (transport time to depot),
	\code{$dTAT} (TurnAround Time at depot), and \code{C} (Purchase cost part)	
}
\source{
    Wagner de Sousa Borges, \emph{Unpublished Book?}, Chapter 4 METRIC example.
    
}
%\references{
%Wagner de Sousa Borges, \emph{Unpublished Book?}, Chapter 4 METRIC example
%}
\keyword{datasets}